requirements.txt
numpy==1.26.4
matplotlib==3.8.3
tensorboard==2.16.2
scipy==1.12.0
seaborn==0.13.2
wandb==0.16.3
moviepy==1.0.3
imageio==2.34.0
e3nn==0.5.1
jsonargparse==4.33.2
jupyter==1.0.0
plotly==5.22.0
pylint == 3.3.1
ase==3.23.0
scikit-learn==1.5.2
optuna==4.5.0
pydantic==2.8.2
PyYAML==6.0.1
ipywidgets==8.0.6
nglview==3.0.6
torch_geometric==2.6.1
torch_scatter==2.1.2
torchmetrics==1.6.1
